requirements.txt
absl-py
actionlib==1.14.0
addict==2.4.0
aiohappyeyeballs
aiohttp
aiosignal
angles==1.9.13
anyio
argon2-cffi
argon2-cffi-bindings
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru
async-timeout
attrs
Babel
backcall==0.2.0
base_local_planner==1.17.3
beautifulsoup4
bleach
blinker
bondpy==1.8.6
Brotli
cachetools
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi
cffi
charset-normalizer
click==8.1.3
cmake==3.26.3
comm==0.1.3
ConfigArgParse==1.5.3
controller-manager==2.40.0
controller-manager-msgs
cryptography
cv-bridge==1.16.2
cycler
dash==2.10.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator
defusedxml
descartes==1.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
docopt==0.6.2
dynamic-reconfigure==1.7.3
entrypoints
et-xmlfile==1.1.0
exceptiongroup
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.3
flatbuffers
frozenlist
gast
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
image-geometry==1.16.2
imageio==2.28.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
interactive-markers==1.12.0
ipykernel==6.23.1
ipython==8.12.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsk_topic_tools==2.2.12
json5
jsonschema==4.23.0
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets==3.0.7
jupyterlab_server
keras==2.13.1
Keras-Preprocessing
kiwisolver
laser_geometry==1.6.7
lazy_loader==0.2
libclang==18.1.1
lit==16.0.5
llvmlite==0.39.1
lxml==4.5.1
Markdown
MarkupSafe==2.1.2
matplotlib
matplotlib-inline==0.1.6
message-filters
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mpmath==1.3.0
multidict
natsort==8.4.0
nbclient
nbconvert
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook
notebook_shim
numba==0.56.4
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib
olefile
open3d==0.17.0
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum
overrides
packaging==23.1
panda==0.3.1
pandas==2.0.3
pandocfilters
parso==0.8.3
pathspec==0.11.1
pexpect
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.4.0
plotly==5.14.1
pooch==1.7.0
progressbar2==4.3.2
prometheus-client
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil
ptyprocess
pure-eval==0.2.2
pyasn1
pyasn1-modules==0.2.8
pycparser
pydot==3.0.1
Pygments==2.15.1
PyJWT
pyOpenSSL
pyparsing
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks
python-dateutil==2.8.2
python-json-logger
python-qt-binding==0.4.4
python-utils==3.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole
QtPy
referencing
requests
requests-oauthlib
resource_retriever==1.12.7
rfc3339-validator
rfc3986-validator
rosbag
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph
roslaunch
roslib==1.15.8
roslint==0.12.0
roslz4
rosmake==1.15.8
rosmaster
rosmsg
rosnode
rosparam
rospy
rosserial_python==0.9.2
rosservice
rostest
rostopic
rosunit==1.15.8
roswtf
rpds-py
rqt-image-view==0.4.17
rqt-moveit
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa
rviz
scikit-image==0.20.0
scikit-learn==0.24.2
scipy==1.9.1
Send2Trash
sensor-msgs==1.13.1
Shapely==1.8.5
six
skan==0.9
smach
smach-ros
smclib==1.8.6
sniffio
sound-play
soupsieve
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
tensorboard
tensorboard-plugin-wit
tensorboard_data_server
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor
terminado
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs
tf2-kdl
tf2-py
tf2-ros
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2
tomli
toolz==0.12.0
topic-tools
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing_extensions
tzdata==2023.3
unidiff==0.7.5
urdfdom-py==0.4.6
urllib3
vcstool==0.3.0
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt
xacro
yamllint==1.31.0
yarg==0.1.9
yarl
zipp==3.15.0
